Move contact type to meta/contact
