+++ src/meta/contact/index.ts
@@ -1,0 +1,2 @@
+export type { Contact, ContactProps, ContactValue } from './contact'
+export { Contacts } from './contacts'
