+++ src/meta/user/index.ts
@@ -1,5 +1,4 @@
 export { Authorizer } from './authorizer'
-export type { Contact, ContactProps } from './contact'
 export type { User } from './user'
 export { UserStatus, UserTitle } from './user'
 export type { UserAuthProvider } from './userAuth'
