+++ src/server/controller/cycleData/nodeExt/createContact.ts
@@ -1,7 +1,8 @@
 import { CountryIso } from 'meta/area'
 import { ActivityLogMessage, Assessment, Cycle } from 'meta/assessment'
+import { Contact } from 'meta/contact'
 import { NodeExt, NodeExtType } from 'meta/nodeExt'
-import { Contact, User } from 'meta/user'
+import { User } from 'meta/user'
 
 import { BaseProtocol, DB } from 'server/db'
 import { NodeExtRepository } from 'server/repository/assessmentCycle/nodeExt'
